[feat] Add common module for reused layers
